model_development/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# MLflow tracking
mlflow>=2.8.0

# Evaluation
evaluate>=0.4.0
sacrebleu>=2.3.0
codebleu>=0.5.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-javascript>=0.20.0
rouge-score>=0.1.2
nltk>=3.8.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Airflow
apache-airflow>=2.7.0
apache-airflow-providers-docker>=3.8.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
